begone code analyzer, i have had enough of you

+++ BossMod/Modules/RealmReborn/Trial/T04PortaDecumana/T04PortaDecumana1.cs
@@ -99,11 +99,11 @@ class T04PortaDecumana1States : StateMachineBuilder
 }
 
 [ModuleInfo(BossModuleInfo.Maturity.Verified, GroupType = BossModuleInfo.GroupType.CFC, GroupID = 830, NameID = 2137, SortOrder = 1)]
-public class T04PortaDecumana1(WorldState ws, Actor primary) : BossModule(ws, primary, new(-772, -600), new ArenaBoundsCircle(19.5f))
+public class T04PortaDecumana1(WorldState ws, Actor primary) : BossModule(ws, primary, new(-772f, -600f), new ArenaBoundsCircle(19.5f))
 {
     protected override void CalculateModuleAIHints(int slot, Actor actor, PartyRolesConfig.Assignment assignment, AIHints hints)
     {
-        if (PrimaryActor.FindStatus(SID.Invincibility) != null || PrimaryActor.FindStatus(SID.VortexBarrier) != null)
+        if (PrimaryActor.FindStatus((uint)SID.Invincibility) != null || PrimaryActor.FindStatus((uint)SID.VortexBarrier) != null)
         {
             var e = hints.FindEnemy(PrimaryActor);
             if (e != null)
